+++ plugins/postcss-relative-color-syntax/CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changes to PostCSS Relative Color Syntax
 
+### Unreleased (patch)
+
+- Fix value calculation for `hue` components in relative color syntax.
+
 ### 1.0.1 (June 1, 2023)
 
 - Updated `@csstools/postcss-progressive-custom-properties` to `2.3.0` (minor)
-
 
 ### 1.0.0 (May 19, 2023)